+++ src/config/FieldConfigs.ts
@@ -18,14 +18,11 @@ export enum InputFieldType {
 	TOGGLE = 'toggle',
 	SLIDER = 'slider',
 	TEXT = 'text',
-	TEXT_AREA_DEPRECATED = 'text_area',
 	TEXT_AREA = 'textArea',
 	SELECT = 'select',
-	MULTI_SELECT_DEPRECATED = 'multi_select',
 	MULTI_SELECT = 'multiSelect',
 	DATE = 'date',
 	TIME = 'time',
-	DATE_PICKER_DEPRECATED = 'date_picker',
 	DATE_PICKER = 'datePicker',
 	NUMBER = 'number',
 	SUGGESTER = 'suggester',
@@ -83,11 +80,6 @@ export const InputFieldConfigs: Record<InputFieldType, InputFieldConfig> = {
 		allowInBlock: true,
 		allowInline: true,
 	},
-	[InputFieldType.TEXT_AREA_DEPRECATED]: {
-		type: InputFieldType.TEXT_AREA_DEPRECATED,
-		allowInBlock: true,
-		allowInline: true,
-	},
 	[InputFieldType.TEXT_AREA]: {
 		type: InputFieldType.TEXT_AREA,
 		allowInBlock: true,
@@ -95,11 +87,6 @@ export const InputFieldConfigs: Record<InputFieldType, InputFieldConfig> = {
 	},
 	[InputFieldType.SELECT]: {
 		type: InputFieldType.SELECT,
-		allowInBlock: true,
-		allowInline: false,
-	},
-	[InputFieldType.MULTI_SELECT_DEPRECATED]: {
-		type: InputFieldType.MULTI_SELECT_DEPRECATED,
 		allowInBlock: true,
 		allowInline: false,
 	},
@@ -115,11 +102,6 @@ export const InputFieldConfigs: Record<InputFieldType, InputFieldConfig> = {
 	},
 	[InputFieldType.TIME]: {
 		type: InputFieldType.TIME,
-		allowInBlock: true,
-		allowInline: true,
-	},
-	[InputFieldType.DATE_PICKER_DEPRECATED]: {
-		type: InputFieldType.DATE_PICKER_DEPRECATED,
 		allowInBlock: true,
 		allowInline: true,
 	},
@@ -311,7 +293,6 @@ export const InputFieldArgumentConfigs: Record<InputFieldArgumentType, InputFiel
 		type: InputFieldArgumentType.OPTION,
 		allowedFieldTypes: [
 			InputFieldType.SELECT,
-			InputFieldType.MULTI_SELECT_DEPRECATED,
 			InputFieldType.MULTI_SELECT,
 			InputFieldType.SUGGESTER,
 			InputFieldType.IMAGE_SUGGESTER,
@@ -366,7 +347,6 @@ export const InputFieldArgumentConfigs: Record<InputFieldArgumentType, InputFiel
 		allowedFieldTypes: [
 			InputFieldType.TEXT,
 			InputFieldType.TEXT_AREA,
-			InputFieldType.TEXT_AREA_DEPRECATED,
 			InputFieldType.NUMBER,
 			InputFieldType.LIST,
 			InputFieldType.INLINE_LIST,