+++ src/utils/errors/ErrorCollectionViewModal.ts
@@ -20,7 +20,7 @@ export class ErrorCollectionViewModal extends Modal {
 	}
 
 	public onOpen(): void {
-		this.modalEl.addClass('mb-error-collection-modal');
+		this.modalEl.addClass('mb-error-collection-modal', 'markdown-rendered');
 		this.titleEl.innerText = 'Meta Bind Error Overview';
 
 		this.component = new ErrorCollectionComponent({